Extract pure goal plan steps
